+++ Block/Adminhtml/Policy/Edit/BackButton.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+
+namespace Panth\RobotsSeo\Block\Adminhtml\Policy\Edit;
+
+use Magento\Framework\View\Element\UiComponent\Control\ButtonProviderInterface;
+
+class BackButton extends GenericButton implements ButtonProviderInterface
+{
+    public function getButtonData(): array
+    {
+        return [
+            'label'      => (string) __('Back'),
+            'on_click'   => sprintf("location.href = '%s';", $this->getUrl('*/*/')),
+            'class'      => 'back',
+            'sort_order' => 10,
+        ];
+    }
+}