+++ Block/Adminhtml/Policy/Edit/SaveButton.php
@@ -1,0 +1,22 @@
+<?php
+declare(strict_types=1);
+
+namespace Panth\RobotsSeo\Block\Adminhtml\Policy\Edit;
+
+use Magento\Framework\View\Element\UiComponent\Control\ButtonProviderInterface;
+
+class SaveButton extends GenericButton implements ButtonProviderInterface
+{
+    public function getButtonData(): array
+    {
+        return [
+            'label'          => (string) __('Save'),
+            'class'          => 'save primary',
+            'data_attribute' => [
+                'mage-init' => ['button' => ['event' => 'save']],
+                'form-role' => 'save',
+            ],
+            'sort_order'     => 90,
+        ];
+    }
+}